Matrix: delegate serial QRCP to strategy method
This commit uses the Strategy pattern to select an appropriate QRCP
pivot computation method, depending on whether the Matrix is
distributed.

The public method qrcp_pivots_transpose will select the appropriate
implementation, and the serial and distributed methods implementing
various QRCP approaches will be private. The distributed version is
not yet implemented.