automake: add GNU boilerplate files/scripts
GNU Automake requires some files be present. There are two rough
categories: scripts and text files.

The following files are text files that are supposed to be
informative:
* AUTHORS
* ChangeLog
* NEWS
* README
* stamp-h1 (file used for timestamp of config header)
* aclocal.m4 (all of the M4 macros needed for configure.ac are stored
  here)

A note about `aclocal.m4`: there is a file of the same name in the
`config` folder. Theoretically, these two files should be the
same. The former is generated when running `autoreconf -I config
--install`, which regenerates all of the build system files in the
correct order; the default output directory for `aclocal` is the
current working directory. Since `aclocal` must be run in the same
directory as the `configure.ac` file, I think the only way to get a
copy of `aclocal.m4` into the `config` directory is to copy it
manually.

The following files are scripts:
* depcomp (computes dependencies among source/object files)
* install-sh (script to install libROM)
* missing (wrapper for missing GNU utilities, e.g., autoconf)