shared_ptr: use C++11 type if avail, else Boost
This commit wraps shared_ptr #include directives and member
declarations in a preprocessor macro that defaults to the C++11
version if the compiler is compiling in C++11 mode (or later), and
falls back to Boost. This construction is helpful for removing Boost
as a dependency of libROM without altering the code too much; Boost is
not installed on the CORAL EA machines (yet).