autoconf: detect Fortran name-mangling convention
This commit adds all the infrastructure needed for detecting the
Fortran name-mangling convention by the Fortran compiler supplied at
the command-line. I tried to make the default Fortran compiler
gfortran, but users may have to prepend FC=$PATH_TO_FORTRAN_COMPILER
to the ./configure command. This kludge is illustrated in the example
configure scripts in the "scripts" directory.