IncrementalSVD: add omitted basis size check
This commit adds a conditional to IncrementalSVD::buildIncrementalSVD
that tests to see if the current number of left singular vectors is
equal to or greater than the dimension of the input vector.

NOTE: d_num_samples appears to equal the number of left singular
vectors in IncrementalSVD and its subclasses -- not the number of
vectors used to build the SVD. This member has different semantics in
StaticSVD and its subclasses, where d_num_samples is actually the
"number of samples", i.e., the number of input vectors used to build
the SVD. This member should probably be renamed.