Add script for rebaselining
This commit adds a script for rebaselining the outputs in the case of
a change, and also outputs diffs. The rebaselined outputs are in the
NEW-BASELINES directory to avoid overwriting the established baselines
in the BASELINES directory. The script also echoes its commands; if
a new baseline differs from the old baseline, a "[ 1 == 1 ]" line
appears in the echoed commands from the script.