IncrementalSVDStandard: fix possible mem leak
This commit fixes a possible memory leak in
IncrementalSVDStandard::computeBasis.

The body of this method currently assigns to d_basis a newly allocated
pointer to a Matrix on each call. If this method is called more than
once, the allocated pointers assigned to d_basis are not freed on
previous calls, and the a new Matrix is allocated and assigned to
d_basis. This method should delete the existing pointer (which is a
no-op if this pointer is NULL) before allocating a new one.