StaticSVD: incl hdr in impl for ScaLAPACK calls
This commit adds the `scalapack_wrapper.h` header in `StaticSVD.C`
because functions from this header are called within
`StaticSVD.C`. Strictly speaking, this inclusion isn't necessary --
it's already included in `StaticSVD.h` -- but adding this explicit
include statement decouples the include statement in `StaticSVD.h`
from the include statement in `StaticSVD.C`, which is desirable
because any of the following scenarios could be true:

- StaticSVD just uses pointers to data types from scalapack_wrapper.h
  and passes them to other functions (so scalapack_wrapper.h would be
  part of an interface library)

- StaticSVD only references data types and functions from
  scalapack_wrapper.h in its implementation (so scalapack_wrapper.h
  would be part of a private library)

- StaticSVD uses data types from scalapack_wrapper.h both in header
  declarations and in its implementation (so scalapack_wrapper.h
  would be part of a public library)

If the data types from scalapack_wrapper.h are no longer used in the
StaticSVD header due to refactoring, deleting the accompanying include
statement would cause a compile-time bug when compiling StaticSVD.C,
which would be annoying to debug and fix. Adding this seemingly
redundant include directive is defensive programming to guard against
future changes, save us some time in future debugging, and to note to
developers that scalapack_wrapper.h is a public dependency in libROM.